+++ tests/FileHelperCallbackTest.php
@@ -7,6 +7,7 @@ namespace Yiisoft\Files\Tests;
 use InvalidArgumentException;
 use Yiisoft\Files\FileHelper;
 use PHPUnit\Framework\TestCase;
+use TypeError;
 
 final class FileHelperCallbackTest extends FileSystemTestCase
 {
@@ -120,7 +121,7 @@ final class FileHelperCallbackTest extends FileSystemTestCase
 
     public function testAfterCopy(): void
     {
-        $compressor = new class () {
+        $compressor = new class() {
             public function compress(string $file): void
             {
                 $mode = 'wb9';
@@ -175,7 +176,7 @@ final class FileHelperCallbackTest extends FileSystemTestCase
 
     public function testCallable(): void
     {
-        $callable = new class ($this) {
+        $callable = new class($this) {
             private TestCase $testCase;
             public bool $beforeCalled = false;
             public bool $afterCalled = false;
@@ -240,7 +241,7 @@ final class FileHelperCallbackTest extends FileSystemTestCase
         $source = $basePath . '/' . $source;
         $destination = $basePath . '/test_dst_dir';
 
-        $this->expectException(InvalidArgumentException::class);
+        $this->expectException(TypeError::class);
 
         FileHelper::copyDirectory($source, $destination, [
             'beforeCopy' => 'not_exists_callback',